refactor: ReviewQueryService 리뷰 요약 메서드 리팩터링

+++ backend/src/main/java/zipgo/review/domain/type/AdverseReactionType.java
@@ -29,7 +29,7 @@ public enum AdverseReactionType {
                 .orElseThrow(AdverseReactionException.NotFound::new);
     }
 
-    public static int getDistributionPercentage(int total, long count) {
+    public static int getDistributionPercentage(long total, long count) {
         return (int) (count * PERCENTAGE / total);
     }
